+++ src/main/python/main.py
@@ -1,0 +1,13 @@
+from fbs_runtime.application_context.PyQt5 import ApplicationContext
+from PyQt5.QtWidgets import QMainWindow
+from src.main.python.gui.mainwindow import MainWindow
+
+import sys
+
+if __name__ == '__main__':
+    appctxt = ApplicationContext()       # 1. Instantiate ApplicationContext
+    window = QMainWindow()
+    window.setCentralWidget(MainWindow())
+    window.show()
+    exit_code = appctxt.app.exec_()      # 2. Invoke appctxt.app.exec_()
+    sys.exit(exit_code)
